Update main.tsx, vars.scss, start.sh, index.tsx, index.tsx, vite.config.ts, WB_Func.tsx, ja.json, zh_CN.json, zh_HK.json, ko.json, es.json, cs.json

+++ sub/web/start.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
-
-if [[ "$LAFTOOLS_ROOT" = "" ]]; then
+set -e
+if [ "$LAFTOOLS_ROOT" = "" ]; then
   echo "LAFTOOLS_ROOT is not set"
   exit 1
 fi
@@ -8,5 +8,5 @@ fi
 cd "$(dirname "$0")"
 
 export FAST_REFRESH=true
-rm -rf ./build
-npm run dev 
+[ -d ./build ] && rm -rf ./build
+FAST_REFRESH=true npm run dev
